+++ webrtc/src/webrtc_c/api/field_trials_view.cc
@@ -1,6 +1,8 @@
 #include "field_trials_view.h"
 
 #include <stddef.h>
+#include <memory>
+#include <string>
 
 // WebRTC
 #include <absl/strings/string_view.h>
@@ -19,5 +21,21 @@ WEBRTC_EXPORT int webrtc_FieldTrialsView_IsEnabled(
     size_t key_len) {
   auto view = reinterpret_cast<const webrtc::FieldTrialsView*>(self);
   return view->IsEnabled(absl::string_view(key, key_len)) ? 1 : 0;
+}
+WEBRTC_EXPORT int webrtc_FieldTrialsView_IsDisabled(
+    const struct webrtc_FieldTrialsView* self,
+    const char* key,
+    size_t key_len) {
+  auto view = reinterpret_cast<const webrtc::FieldTrialsView*>(self);
+  return view->IsDisabled(absl::string_view(key, key_len)) ? 1 : 0;
+}
+WEBRTC_EXPORT struct std_string_unique* webrtc_FieldTrialsView_Lookup(
+    const struct webrtc_FieldTrialsView* self,
+    const char* key,
+    size_t key_len) {
+  auto view = reinterpret_cast<const webrtc::FieldTrialsView*>(self);
+  auto value = std::make_unique<std::string>(
+      view->Lookup(absl::string_view(key, key_len)));
+  return reinterpret_cast<struct std_string_unique*>(value.release());
 }
 }